+++ Pokemon/Calculators/TempStatsCalculator.cs
@@ -5,7 +5,7 @@ namespace Pokemon.Calculators
 {
     public static class TempStatsCalculator
     {
-        public static int GetAttack(IPokemon pokemon)
+        public static float GetAttack(IPokemon pokemon)
         {
             float stageMultipler = StageHelper.StageToMultipler(pokemon.StatModifierStages[(int)StatType.Attack]);
             int conditionMultipler = pokemon.Condition != Condition.BRN ? 1 : 2;
@@ -13,28 +13,28 @@ namespace Pokemon.Calculators
             return attack;
         }
 
-        public static int GetDefence(IPokemon pokemon)
+        public static float GetDefence(IPokemon pokemon)
         {
             float stageMultipler = StageHelper.StageToMultipler(pokemon.StatModifierStages[(int)StatType.Defence]);
             int defence = Convert.ToInt32(pokemon.Stats.Defence * stageMultipler);
             return defence;
         }
 
-        public static int GetSpecialAttack(IPokemon pokemon)
+        public static float GetSpecialAttack(IPokemon pokemon)
         {
             float stageMultipler = StageHelper.StageToMultipler(pokemon.StatModifierStages[(int)StatType.SpecialAttack]);
             int specialAttack = Convert.ToInt32(pokemon.Stats.SpecialAttack * stageMultipler);
             return specialAttack;
         }
 
-        public static int GetSpecialDefence(IPokemon pokemon)
+        public static float GetSpecialDefence(IPokemon pokemon)
         {
             float stageMultipler = StageHelper.StageToMultipler(pokemon.StatModifierStages[(int)StatType.SpecialDefence]);
             int specialDefence = Convert.ToInt32(pokemon.Stats.SpecialDefence * stageMultipler);
             return specialDefence;
         }
 
-        public static int GetSpeed(IPokemon pokemon)
+        public static float GetSpeed(IPokemon pokemon)
         {
             float stageMultipler = StageHelper.StageToMultipler(pokemon.StatModifierStages[(int)StatType.Speed]);
             int conditionMultipler = Convert.ToInt32(pokemon.Condition != Condition.PAR ? 1 : 1.5f);
